moved cosine_dict to a file; main() now uses cosine_dict() & embeddings() functions
